clean code

+++ PoliNetworkBot_CSharp/Code/Utils/Logger.cs
@@ -1,8 +1,5 @@
-﻿using JsonPolimi_Core_nf.Tipi;
-using PoliNetworkBot_CSharp.Code.Bots.Anon;
-using PoliNetworkBot_CSharp.Code.Enums;
-using PoliNetworkBot_CSharp.Code.Objects;
-using PoliNetworkBot_CSharp.Code.Objects.TelegramMedia;
+﻿#region
+
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -11,30 +8,39 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Threading.Tasks.Dataflow;
+using System.Web;
+using JsonPolimi_Core_nf.Tipi;
+using PoliNetworkBot_CSharp.Code.Bots.Anon;
+using PoliNetworkBot_CSharp.Code.Enums;
+using PoliNetworkBot_CSharp.Code.Objects;
+using PoliNetworkBot_CSharp.Code.Objects.TelegramMedia;
 using Telegram.Bot.Types.Enums;
 using TeleSharp.TL;
+
+#endregion
 
 namespace PoliNetworkBot_CSharp.Code.Utils
 {
     public static class Logger
     {
-        private static readonly Dictionary<long, TelegramBotAbstract> Subscribers = new();
-        private static readonly BufferBlock<MessageQueue> Buffer = new();
-        private static readonly Object Lock = new();
         private const string DataLogPath = "./data/log.txt";
         private const string LogSeparator = "#@#LOG ENTRY#@#";
+        private static readonly Dictionary<long, TelegramBotAbstract> Subscribers = new();
+        private static readonly BufferBlock<MessageQueue> Buffer = new();
+        private static readonly object Lock = new();
+
+        private static readonly object PrintLogLock = new();
 
         internal static async Task MainMethodAsync()
         {
             while (await Buffer.OutputAvailableAsync())
-            {
                 try
                 {
                     var messageToBeSent = Buffer.Receive();
-                    string escaped = System.Web.HttpUtility.HtmlEncode(messageToBeSent.text);
+                    var escaped = HttpUtility.HtmlEncode(messageToBeSent.text);
                     var text = new Language(new Dictionary<string, string>
                     {
-                        {"un", escaped}
+                        { "un", escaped }
                     });
                     await messageToBeSent.key.Value.SendTextMessageAsync(messageToBeSent.key.Key, text,
                         messageToBeSent.ChatType, "un", ParseMode.Html,
@@ -47,7 +53,6 @@ namespace PoliNetworkBot_CSharp.Code.Utils
                     Console.WriteLine(e);
                     Console.WriteLine("##########################");
                 }
-            }
         }
 
         public static void WriteLine(object log, LogSeverityLevel logSeverityLevel = LogSeverityLevel.INFO)
@@ -57,30 +62,26 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             try
             {
                 Console.WriteLine(logSeverityLevel + " | " + log);
-                string log1 = log.ToString();
-                if (Directory.Exists("./data/") == false)
-                {
-                    Directory.CreateDirectory("./data/");
-                }
+                var log1 = log.ToString();
+                if (Directory.Exists("./data/") == false) Directory.CreateDirectory("./data/");
 
-                if (!File.Exists(DataLogPath))
-                {
-                    File.WriteAllText(DataLogPath, "");
-                }
+                if (!File.Exists(DataLogPath)) File.WriteAllText(DataLogPath, "");
                 lock (Lock)
                 {
-                    File.AppendAllLinesAsync(DataLogPath, new[] { "#@#LOG ENTRY#@#" +  GetTime()
-                        + " | " + logSeverityLevel + " | " + log1 });
+                    File.AppendAllLinesAsync(DataLogPath, new[]
+                    {
+                        "#@#LOG ENTRY#@#" + GetTime()
+                                          + " | " + logSeverityLevel + " | " + log1
+                    });
                 }
+
                 foreach (var subscriber in Subscribers)
-                {
                     Buffer.Post(
                         new MessageQueue(subscriber,
                             log1,
                             ChatType.Group,
                             ParseMode.Html)
                     );
-                }
             }
             catch (Exception e)
             {
@@ -96,7 +97,8 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             return DateTime.UtcNow.ToString("yyyy-MM-dd HH:mm:ss.fff", CultureInfo.InvariantCulture);
         }
 
-        public static async Task Subscribe(long fromId, TelegramBotAbstract telegramBotAbstract, MessageEventArgs messageEventArgs)
+        public static async Task Subscribe(long fromId, TelegramBotAbstract telegramBotAbstract,
+            MessageEventArgs messageEventArgs)
         {
             try
             {
@@ -123,8 +125,6 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             }
         }
 
-        private static readonly object PrintLogLock = new();
-
         public static void PrintLog(TelegramBotAbstract sender, List<long> sendTo, MessageEventArgs messageEventArgs)
         {
             lock (PrintLogLock)
@@ -140,17 +140,13 @@ namespace PoliNetworkBot_CSharp.Code.Utils
                     }
                     catch (Exception e)
                     {
-                        Logger.WriteLine(e, LogSeverityLevel.CRITICAL);
+                        WriteLine(e, LogSeverityLevel.CRITICAL);
                     }
 
                     if (text is { Count: <= 1 })
-                    {
                         EmptyLog(sender, sendTo);
-                    }
                     else
-                    {
                         PrintLog2(sendTo, sender, path);
-                    }
                 }
                 catch (Exception e)
                 {
@@ -161,14 +157,14 @@ namespace PoliNetworkBot_CSharp.Code.Utils
 
         private static void PrintLog2(List<long> sendTo, TelegramBotAbstract sender, string path)
         {
-            string file = File.ReadAllText(path);
+            var file = File.ReadAllText(path);
             file = string.Join("", file.Split(LogSeparator)); //remove "#@#LOG ENTRY#@#" from all the lines
-            Encoding encoding = Encoding.UTF8;
+            var encoding = Encoding.UTF8;
 
             var text2 = new Language(new Dictionary<string, string>
-                    {
-                        {"it", "LOG:"}
-                    });
+            {
+                { "it", "LOG:" }
+            });
 
             foreach (var sendToSingle in sendTo)
             {
@@ -189,15 +185,43 @@ namespace PoliNetworkBot_CSharp.Code.Utils
         private static void EmptyLog(TelegramBotAbstract sender, List<long> sendTo)
         {
             var text = new Language(new Dictionary<string, string>
-                    {
-                        {"en", "No log available."}
-                    });
+            {
+                { "en", "No log available." }
+            });
 
             foreach (var sendToSingle in sendTo)
-            {
                 SendMessage.SendMessageInPrivate(sender, sendToSingle, "en",
                     null, text, ParseMode.Html, null).Wait();
-            }
+        }
+
+        internal static void Log(EventoConLog eventoLog)
+        {
+            var log = eventoLog.GetLog();
+            for (var i = 0; i < log.Item2; i++) WriteLine(log.Item1[i]);
+        }
+
+        public static bool ContainsCriticalErrors(out string s)
+        {
+            var log = File.ReadAllText(DataLogPath);
+            var entries = log.Split(LogSeparator).ToList();
+            s = "";
+            var toReturn = false;
+            foreach (var entry in entries)
+                try
+                {
+                    var severityLevel = entry[(GetTime().Length + 3)..];
+                    if (!severityLevel.StartsWith(LogSeverityLevel.NOTICE.ToString()) &&
+                        !severityLevel.StartsWith(LogSeverityLevel.WARNING.ToString()) &&
+                        !severityLevel.StartsWith(LogSeverityLevel.CRITICAL.ToString()) &&
+                        !severityLevel.StartsWith(LogSeverityLevel.EMERGENCY.ToString())) continue;
+                    s += entry;
+                    toReturn = true;
+                }
+                catch (ArgumentOutOfRangeException)
+                {
+                }
+
+            return toReturn;
         }
 
         public static class GroupsFixLog
@@ -206,8 +230,8 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             private static Dictionary<long, KeyValuePair<string, Exception>> _newNull = new();
             private static List<string> _nameChange = new();
             private static List<string> _oldNull = new();
-            private static int _countFixed = 0;
-            private static int _countIgnored = 0;
+            private static int _countFixed;
+            private static int _countIgnored;
 
             public static void SendLog(TelegramBotAbstract telegramBotAbstract, MessageEventArgs messageEventArgs,
                 GroupsFixLogUpdatedEnum groupsFixLogUpdatedEnum = GroupsFixLogUpdatedEnum.ALL)
@@ -218,7 +242,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils
                 {
                     message += "-NewTitle null && OldTitle null: [Group ID]";
                     message += "\n";
-                    message = _bothNull.Aggregate(message, (current, @group) => current + @group + "\n");
+                    message = _bothNull.Aggregate(message, (current, group) => current + group + "\n");
                     message += "\n\n";
                     message += "-NewTitle null:";
                     message += "\n";
@@ -229,16 +253,17 @@ namespace PoliNetworkBot_CSharp.Code.Utils
                     message = _oldNull.Aggregate(message, (current, newTitle) => current + newTitle + "\n");
                     message += "\n\n";
                 }
+
                 message += "-Name Changed: [oldTitle [->] newTitle]";
                 message += "\n";
-                message = _nameChange.Aggregate(message, (current, nameChange) => current + (nameChange + "\n"));
+                message = _nameChange.Aggregate(message, (current, nameChange) => current + nameChange + "\n");
                 message += "\n\n";
                 message += "Fixed: " + _countFixed;
                 message += "\nIgnored (already ok): " + _countIgnored;
 
-                string escaped = System.Web.HttpUtility.HtmlEncode(message);
-                Logger.WriteLine(message);
-                Utils.NotifyUtil.NotifyOwners(escaped, telegramBotAbstract, messageEventArgs);
+                var escaped = HttpUtility.HtmlEncode(message);
+                WriteLine(message);
+                NotifyUtil.NotifyOwners(escaped, telegramBotAbstract, messageEventArgs);
                 Reset();
             }
 
@@ -256,21 +281,17 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             {
                 var toReturn = "";
                 var exceptionTypes = new List<Type>();
-                foreach (var (_, exception) in newNull.Values)
-                {
-                    if (!exceptionTypes.Contains(exception.GetType()))
-                        exceptionTypes.Add(exception.GetType());
-                }
+                foreach (var exception in newNull.Values.Cast<Exception>()
+                             .Where(exception => !exceptionTypes.Contains(exception.GetType())))
+                    exceptionTypes.Add(exception.GetType());
 
                 foreach (var exceptionType in exceptionTypes)
                 {
                     toReturn += "#" + exceptionType + ":\n";
                     toReturn += "[GroupId , oldTitle]\n";
-                    foreach (var group in newNull)
-                    {
-                        if (newNull[group.Key].Value.GetType() == exceptionType)
-                            toReturn += group.Key + " , " + group.Value.Key + "\n";
-                    }
+                    toReturn = newNull.Where(group => newNull[group.Key].Value.GetType() == exceptionType)
+                        .Aggregate(toReturn,
+                            (current, group) => current + @group.Key + " , " + @group.Value.Key + "\n");
                 }
 
                 return toReturn;
@@ -283,6 +304,7 @@ namespace PoliNetworkBot_CSharp.Code.Utils
                     _bothNull.Add(tableRowId + " but GetChat responded with ID: " + item1Id);
                     return;
                 }
+
                 _bothNull.Add(tableRowId.ToString());
             }
 
@@ -306,43 +328,6 @@ namespace PoliNetworkBot_CSharp.Code.Utils
             {
                 _countIgnored++;
             }
-        }
-
-        internal static void Log(EventoConLog eventoLog)
-        {
-            var log = eventoLog.GetLog();
-            for (int i = 0; i < log.Item2; i++)
-            {
-                Logger.WriteLine(log.Item1[i]);
-            }
-        }
-
-        public static bool ContainsCriticalErrors(out string s)
-        {
-            var log = File.ReadAllText(DataLogPath);
-            var entries = log.Split(LogSeparator).ToList();
-            s = "";
-            var toReturn = false;
-            foreach (var entry in entries)
-            {
-                try
-                {
-                    var severityLevel = entry[(GetTime().Length + 3)..];
-                    if (severityLevel.StartsWith(LogSeverityLevel.NOTICE.ToString())
-                        || severityLevel.StartsWith(LogSeverityLevel.WARNING.ToString())
-                        || severityLevel.StartsWith(LogSeverityLevel.CRITICAL.ToString())
-                        || severityLevel.StartsWith(LogSeverityLevel.EMERGENCY.ToString())
-                       )
-                    {
-                        s += entry;
-                        toReturn = true;
-                    }
-                }
-                catch (ArgumentOutOfRangeException)
-                {
-                }
-            }
-            return toReturn;
         }
     }
 }
